updated env variable name
